# docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "XAPI endpoints for the XNAT Container Service",
    "version" : "1.4.0-SNAPSHOT",
    "title" : "Container Service API",
    "contact" : {
      "name" : "John Flavin",
      "email" : "jflavin@wustl.edu"
    },
    "license" : {
      "name" : "BSD 2-Clause",
      "url" : "https://github.com/NrgXnat/container-service/blob/master/LICENSE.md"
    }
  },
  "host" : "xnat.org",
  "basePath" : "/xapi",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/docker/hubs" : {
      "get" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Get Docker Hubs",
        "description" : "",
        "operationId" : "getHubs",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DockerHubWithPing"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Create new Docker Hub",
        "description" : "",
        "operationId" : "createHub",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DockerHub"
          }
        }, {
          "name" : "default",
          "in" : "query",
          "required" : true,
          "type" : "boolean"
        }, {
          "name" : "reason",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DockerHubWithPing"
            }
          }
        }
      }
    },
    "/docker/hubs/{id}" : {
      "get" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Get Docker Hub by ID",
        "description" : "",
        "operationId" : "getHub",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DockerHubWithPing"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Update Docker Hub by ID",
        "description" : "",
        "operationId" : "updateHub",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DockerHub"
          }
        }, {
          "name" : "default",
          "in" : "query",
          "required" : true,
          "type" : "boolean"
        }, {
          "name" : "reason",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Delete Docker Hub by ID",
        "description" : "",
        "operationId" : "deleteHub",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/docker/hubs/{id}/pull" : {
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Pull image from Docker Hub by ID",
        "description" : "",
        "operationId" : "pullImageFromHub",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        }, {
          "name" : "image",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "save-commands",
          "in" : "query",
          "required" : true,
          "type" : "boolean"
        }, {
          "name" : "username",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/docker/hubs/{name}" : {
      "get" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Get Docker Hub by Name",
        "description" : "",
        "operationId" : "getHub",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d*[^\\d]+\\d*"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DockerHubWithPing"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Delete Docker Hub by Name",
        "description" : "",
        "operationId" : "deleteHub",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d*[^\\d]+\\d*"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/docker/hubs/{name}/pull" : {
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Pull image from Docker Hub by Name",
        "description" : "",
        "operationId" : "pullImageFromHub",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d*[^\\d]+\\d*"
        }, {
          "name" : "image",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "save-commands",
          "in" : "query",
          "required" : true,
          "type" : "boolean"
        }, {
          "name" : "username",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/docker/images" : {
      "get" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Get list of images.",
        "description" : "Returns a list of all Docker images on the Docker server.",
        "operationId" : "getImages",
        "produces" : [ "application/json;charset=UTF-8" ],
        "responses" : {
          "200" : {
            "description" : "A list of images on the server",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DockerImage"
              }
            }
          },
          "424" : {
            "description" : "Admin must set up Docker server."
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
    "/docker/images/save" : {
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Save Commands from labels",
        "description" : "Read labels from Docker image. If any labels contain key org.nrg.commands, parse value as list of Commands.",
        "operationId" : "saveFromLabels",
        "parameters" : [ {
          "name" : "image",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Command"
              }
            }
          }
        }
      }
    },
    "/docker/images/{id}" : {
      "get" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Get Docker image",
        "description" : "Retrieve information about a Docker image from the docker server",
        "operationId" : "getImage",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Found the image",
            "schema" : {
              "$ref" : "#/definitions/DockerImage"
            }
          },
          "404" : {
            "description" : "No docker image with given id on the server"
          },
          "424" : {
            "description" : "Admin must set up Docker server."
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "delete" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Delete Docker image",
        "description" : "Remove information about a Docker image",
        "operationId" : "deleteImage",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "force",
          "in" : "query",
          "required" : true,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "Image was removed"
          },
          "404" : {
            "description" : "No docker image with given id on docker server"
          },
          "424" : {
            "description" : "Admin must set up Docker server."
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
    "/docker/pull" : {
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Pull image from default Docker Hub",
        "description" : "",
        "operationId" : "pullImageFromDefaultHub",
        "parameters" : [ {
          "name" : "image",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "save-commands",
          "in" : "query",
          "required" : true,
          "type" : "boolean"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/docker/server" : {
      "get" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Docker server",
        "description" : "Returns Docker server configuration values",
        "operationId" : "getServer",
        "produces" : [ "application/json;charset=UTF-8" ],
        "responses" : {
          "200" : {
            "description" : "The Docker server configuration",
            "schema" : {
              "$ref" : "#/definitions/DockerServer"
            }
          },
          "400" : {
            "description" : "The server has not been configured"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "post" : {
        "tags" : [ "Docker API for XNAT Container Service" ],
        "summary" : "Set Docker server configuration",
        "description" : "Save new Docker server configuration values",
        "operationId" : "setServer",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DockerServer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DockerServerWithPing"
            }
          },
          "201" : {
            "description" : "The Docker server configuration was saved"
          },
          "400" : {
            "description" : "Must set the \"host\" property in request body"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Command" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "name" : {
          "type" : "string",
          "readOnly" : true
        },
        "label" : {
          "type" : "string",
          "readOnly" : true
        },
        "description" : {
          "type" : "string",
          "readOnly" : true
        },
        "version" : {
          "type" : "string",
          "readOnly" : true
        },
        "schema-version" : {
          "type" : "string",
          "readOnly" : true
        },
        "info-url" : {
          "type" : "string",
          "readOnly" : true
        },
        "image" : {
          "type" : "string",
          "readOnly" : true
        },
        "type" : {
          "type" : "string",
          "readOnly" : true
        },
        "index" : {
          "type" : "string",
          "readOnly" : true
        },
        "hash" : {
          "type" : "string",
          "readOnly" : true
        },
        "working-directory" : {
          "type" : "string",
          "readOnly" : true
        },
        "command-line" : {
          "type" : "string",
          "readOnly" : true
        },
        "mounts" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/CommandMount"
          }
        },
        "environment-variables" : {
          "type" : "object",
          "readOnly" : true,
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "ports" : {
          "type" : "object",
          "readOnly" : true,
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "inputs" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/CommandInput"
          }
        },
        "outputs" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/CommandOutput"
          }
        },
        "xnat" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/CommandWrapper"
          }
        }
      }
    },
...
    "DockerHub" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "name" : {
          "type" : "string",
          "readOnly" : true
        },
        "url" : {
          "type" : "string",
          "readOnly" : true
        },
        "default" : {
          "type" : "boolean",
          "readOnly" : true
        }
      }
    },
    "DockerHubWithPing" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "name" : {
          "type" : "string",
          "readOnly" : true
        },
        "url" : {
          "type" : "string",
          "readOnly" : true
        },
        "default" : {
          "type" : "boolean",
          "readOnly" : true
        },
        "ping" : {
          "type" : "boolean",
          "readOnly" : true
        }
      }
    },
    "DockerImage" : {
      "type" : "object",
      "properties" : {
        "image-id" : {
          "type" : "string",
          "readOnly" : true
        },
        "tags" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "labels" : {
          "type" : "object",
          "readOnly" : true,
          "additionalProperties" : {
            "type" : "string"
          }
        }
      },
      "description" : "Properties that define an image."
    },
...
  }
}